+++ src/js/components/internal/slider-transitioner.js
@@ -1,5 +1,5 @@
 import {translate} from '../../mixin/internal/slideshow-animations';
-import {children, clamp, createEvent, css, Deferred, dimensions, includes, index, isRtl, noop, position, sortBy, Transition, trigger} from 'uikit-util';
+import {children, clamp, createEvent, css, Deferred, dimensions, findIndex, includes, isRtl, noop, position, Transition, trigger} from 'uikit-util';
 
 export default function (prev, next, dir, {center, easing, list}) {
 
@@ -23,7 +23,6 @@ export default function (prev, next, dir, {center, easing, list}) {
 
             this.translate(percent);
 
-            prev && this.updateTranslates();
             percent = prev ? percent : clamp(percent, 0, 1);
             triggerUpdate(this.getItemIn(), 'itemin', {percent, duration, timing, dir});
             prev && triggerUpdate(this.getItemIn(true), 'itemout', {percent: 1 - percent, duration, timing, dir});
@@ -59,13 +58,29 @@ export default function (prev, next, dir, {center, easing, list}) {
                 dimensions(list).width
             ) * (isRtl ? -1 : 1), 'px'));
 
-            this.updateTranslates();
+            const actives = this.getActives();
+            const itemIn = this.getItemIn();
+            const itemOut = this.getItemIn(true);
 
-            if (prev) {
-                percent = clamp(percent, -1, 1);
-                triggerUpdate(this.getItemIn(), 'itemtranslatein', {percent, dir});
-                triggerUpdate(this.getItemIn(true), 'itemtranslateout', {percent: 1 - percent, dir});
-            }
+            percent = prev ? clamp(percent, -1, 1) : 0;
+
+            children(list).forEach((slide, i) => {
+                const isActive = includes(actives, slide);
+                const isIn = slide === itemIn;
+                const isOut = slide === itemOut;
+                const translateIn = isIn || !isOut && (isActive || dir * (isRtl ? -1 : 1) === -1 ^ getElLeft(slide, list) > getElLeft(prev || next));
+
+                triggerUpdate(slide, `itemtranslate${translateIn ? 'in' : 'out'}`, {
+                    dir,
+                    percent: isOut
+                        ? 1 - percent
+                        : isIn
+                            ? percent
+                            : isActive
+                                ? 1
+                                : 0
+                });
+            });
 
         },
 
@@ -79,33 +94,21 @@ export default function (prev, next, dir, {center, easing, list}) {
 
         getItemIn(out = false) {
 
-            const actives = sortBy(this.getActives(), 'offsetLeft');
-            const all = sortBy(children(list), 'offsetLeft');
-            const i = index(all, actives[dir * (out ? -1 : 1) > 0 ? actives.length - 1 : 0]);
+            let actives = this.getActives();
+            let nextActives = inView(list, getLeft(next || prev, list, center));
 
-            return ~i && all[i + (prev && !out ? dir : 0)];
+            if (out) {
+                const temp = actives;
+                actives = nextActives;
+                nextActives = temp;
+            }
+
+            return nextActives[findIndex(nextActives, el => !includes(actives, el))];
 
         },
 
         getActives() {
-            return [prev || next].concat(children(list).filter(slide => {
-                const slideLeft = getElLeft(slide, list);
-                return slideLeft > from && slideLeft + dimensions(slide).width <= dimensions(list).width + from;
-            }));
-        },
-
-        updateTranslates() {
-
-            const actives = this.getActives();
-
-            children(list).forEach(slide => {
-                const isActive = includes(actives, slide);
-
-                triggerUpdate(slide, `itemtranslate${isActive ? 'in' : 'out'}`, {
-                    percent: isActive ? 1 : 0,
-                    dir: slide.offsetLeft <= next.offsetLeft ? 1 : -1
-                });
-            });
+            return inView(list, getLeft(prev || next, list, center));
         }
 
     };
@@ -136,6 +139,19 @@ function centerEl(el, list) {
 
 export function getElLeft(el, list) {
     return el && (position(el).left + (isRtl ? dimensions(el).width - dimensions(list).width : 0)) * (isRtl ? -1 : 1) || 0;
+}
+
+function inView(list, listLeft) {
+
+    listLeft -= 1;
+    const listRight = listLeft + dimensions(list).width + 2;
+
+    return children(list).filter(slide => {
+        const slideLeft = getElLeft(slide, list);
+        const slideRight = slideLeft + dimensions(slide).width;
+
+        return slideLeft >= listLeft && slideRight <= listRight;
+    });
 }
 
 function triggerUpdate(el, type, data) {